+++ packages/base-controller/src/next/BaseController.ts
@@ -367,6 +367,7 @@ export class BaseController<
  * @param state - The full controller state.
  * @param metadata - The controller metadata.
  * @param metadataProperty - The metadata property to use to derive state.
+ * @param captureException - Reports an error to an error monitoring service.
  * @returns The metadata-derived controller state.
  */
 export function deriveStateFromMetadata<
@@ -375,6 +376,7 @@ export function deriveStateFromMetadata<
   state: ControllerState,
   metadata: StateMetadata<ControllerState>,
   metadataProperty: keyof StatePropertyMetadata<Json>,
+  captureException?: (error: Error) => void,
 ): Record<keyof ControllerState, Json> {
   return (Object.keys(state) as (keyof ControllerState)[]).reduce<
     Record<keyof ControllerState, Json>
@@ -393,11 +395,23 @@ export function deriveStateFromMetadata<
       }
       return derivedState;
     } catch (error) {
-      // Throw error after timeout so that it is captured as a console error
-      // (and by Sentry) without interrupting state-related operations
-      setTimeout(() => {
-        throw error;
-      });
+      // Capture error without interrupting state-related operations
+      // See [ADR core#0016](https://github.com/MetaMask/decisions/blob/main/decisions/core/0016-core-classes-error-reporting.md)
+      if (captureException) {
+        try {
+          captureException(
+            error instanceof Error ? error : new Error(String(error)),
+          );
+        } catch (captureExceptionError) {
+          console.error(
+            new Error(`Error thrown when calling 'captureException'`),
+            captureExceptionError,
+          );
+          console.error(error);
+        }
+      } else {
+        console.error(error);
+      }
       return derivedState;
     }
   }, {} as never);